requirements.txt
streamlit
Pillow
surya-ocr
torch
transformers
tiktoken
torchvision
verovio
accelerate
rapidfuzz
gradio
